# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rank API Documentation",
    "description": "API documentation for Rank - SEO and content management platform",
    "version": "0.1.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://rank.brnd.dev",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/user/account/deletion": {
      "get": {
        "tags": ["Account"],
        "summary": "Get deletion status",
        "description": "Get current deletion status and summary",
        "operationId": "getDeletionStatus",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Deletion status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Account"],
        "summary": "Request account deletion",
        "description": "Request or confirm account deletion",
        "operationId": "requestAccountDeletion",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deletion request created or confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "message": {"type": "string"},
                    "confirmationToken": {"type": "string"}
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "delete": {
        "tags": ["Account"],
        "summary": "Cancel deletion request",
        "description": "Cancel a pending deletion request",
        "operationId": "cancelDeletionRequest",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deletion request cancelled"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk JWT authentication token"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {"type": "string"},
          "message": {"type": "string"}
        }
      },
...
      "DeletionStatusResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "data": {
            "type": "object",
            "properties": {
              "hasPendingRequest": {"type": "boolean"},
              "pendingRequest": {"type": "object"},
              "summary": {"type": "object"},
              "activeSubscriptions": {"type": "array"}
            }
          }
        }
      },
      "AccountDeletionRequest": {
        "type": "object",
        "properties": {
          "confirmationToken": {"type": "string"},
          "confirmed": {"type": "boolean"}
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}